Update 2016_08_02_183143_add_slug_field_for_discussions.php
Hi 

When I run tests in sqlite or use sqlite DB, it gives error in running migration. I am attaching a laracast link  for it. That explains it. https://laracasts.com/discuss/channels/general-discussion/migrations-sqlite-general-error-1-cannot-add-a-not-null-column-with-default-value-null
Please update migration. So it can pass tests or will be used with  sqlite DB.

TIA